% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_canopy_map.R
\name{base_map}
\alias{base_map}
\title{Maps canopy cover for structure}
\usage{
base_map()
}
\arguments{
\item{df}{sf object}

\item{structureIsSuburb}{TRUE/FALSE}

\item{structureID}{Unique identifier for structure}

\item{structureName}{Name for structure polygon}

\item{structureTree}{Canopy cover for structures}

\item{minZoom}{Minimum map zoom level (1-12)}

\item{maxZoom}{Maximum map zoom level (1-12)}

\item{lng1}{Longitude of top left bounding box}

\item{lat1}{Latitude of top left bounding box}

\item{lng2}{Longitude of bottom right bounding box}

\item{lat2}{Latitude of bottom right bounding box}

\item{viewZoom}{}
}
\value{
leaflet map of structure with tree canopy symbology
}
\description{
Maps canopy cover for structure
}
