% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_echarts.R
\name{echarts_vegtype_pie}
\alias{echarts_vegtype_pie}
\title{echarts4r plots for geometric structure}
\usage{
echarts_vegtype_pie(df, values, columnNames)
}
\value{
A echarts4r plots
}
\description{
echarts4r plots for geometric structure
}
