% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_piechart.R
\name{filter_SSC}
\alias{filter_SSC}
\title{Filters sf object to return piechart data from map click}
\usage{
filter_SSC(df, uniqueID, clickID)
}
\arguments{
\item{df}{sf object}

\item{uniqueID}{Unique identifier for geometric structure, must match id of leaflet click event}

\item{columnsToDrop}{drops unnecessary columns from df}

\item{tree}{percentage tree in geometric structure}

\item{shrub}{percentage grass in geometric structure}

\item{grass}{percentage shrub in geometric structure}
}
\value{
Filtered data ready for piechart plotting
}
\description{
Filters sf object to return piechart data from map click
}
