% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_LGA_map.R
\name{LGA_map}
\alias{LGA_map}
\title{Generates tree canopy map for LGA structures}
\usage{
LGA_map(df, uniqueID, lng1, lat1, lng2, lat2)
}
\arguments{
\item{df}{sf object}

\item{uniqueID}{A unique identifier for each polygon that will be return during the click event}

\item{lng1}{Longitude of top left bounding box}

\item{lat1}{Latitude of top left bounding box}

\item{lng2}{Longitude of bottom right bounding box}

\item{lat2}{Latitude of bottom right bounding box}
}
\value{
leaflet map of LGA structure with tree canopy symbology
}
\description{
Generates tree canopy map for LGA structures
}
