% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_labels.R
\name{map_labels}
\alias{map_labels}
\title{Generate map labels for structure canopy cover}
\usage{
map_labels(labelName, labelTree, labelShrub, labelGrass)
}
\arguments{
\item{structureName}{The name of the geometric structure e.g. suburb name}

\item{structurePercent}{Tree canopy cover of structure}
}
\value{
Labels to be passed into a leaflet map
}
\description{
Generate map labels for structure canopy cover
}
